Reduce compile warnings 2:
use of C++ casts in cases where data types were being implicitly
converted from larger to smaller data types

+++ src/Skyrim/VMAD/Alias.cpp
@@ -109,7 +109,7 @@ void Alias::Write(FileWriter &writer)
     writer.record_write(&version, sizeof(version));
     int16_t objFormat = 2;
     writer.record_write(&objFormat, sizeof(objFormat));
-    uint16_t count = scripts.size();
+    uint16_t count = static_cast<uint16_t>(scripts.size());
     writer.record_write(&count, sizeof(count));
     for (uint16_t i = 0; i < count; ++i)
         scripts[i]->Write(writer);
